sheetpost v1

+++ src/API/UnsheetPost.php
@@ -3,7 +3,6 @@
 namespace Sheetpost\API;
 
 use Exception;
-use JetBrains\PhpStorm\ArrayShape;
 
 class UnsheetPost extends Response
 {
@@ -16,19 +15,21 @@ class UnsheetPost extends Response
     /**
      * @throws Exception
      */
-    #[ArrayShape(["success" => "bool"])]
     protected function getQueryResponse(array $getParameters): array
     {
         $username = $getParameters['username'];
         $password = $getParameters['password'];
+        $postId = $getParameters['post_id'];
+        if (!ctype_digit($postId)) {
+            return ['success' => false, 'error' => 'post id is not an integer'];
+        }
         if ($this->db->isUserExists($username, $password)) {
-            $postId = $getParameters['post_id'];
             $this->db->query("
                 DELETE FROM sheets
                     WHERE username='$username' AND post_id=$postId"
             );
             return ['success' => true];
         }
-        return ['success' => false];
+        return ['success' => false, 'error' => 'invalid username or password'];
     }
 }
